Axe added and new art for sword
Enemies and player have separate paint jobs for their weapons. Axe has been added (is a slower but higher damage weapon). Also, basic particles added for hits

+++ Lifetide/Assets/_Scripts/Ui/UIUpdater.cs
@@ -42,7 +42,7 @@ public class UIUpdater : MonoBehaviour
             if (thisInfoStore.CheckInfoLock(InfoType.Health))
             {
                 thisInfoStore.TryGetInfo(InfoType.Health, out float health);
-                healthTXT.text = "Health: " + health;
+                healthTXT.text = "Health: " + (int)health;
             }
 
             // Check and update shield display
@@ -56,7 +56,7 @@ public class UIUpdater : MonoBehaviour
             if (thisInfoStore.CheckInfoLock(InfoType.Dashes))
             {
                 thisInfoStore.TryGetInfo(InfoType.Dashes, out float dashes);
-                dashesTXT.text = "Dashes: " + dashes;
+                dashesTXT.text = "Dashes: " + (int)dashes;
             }
         }
     }
